Add support for global constructor/destructor
In order to call constructor and destructors of global objects, the
kernel have to bind some procedures to call at program startup and
ending. This is done with 4 objects files that has to be linked in a
specific order, around all other objects files.

These files are crtbegin.o, crtend.o, crti.o and crtn.o. The two
formers are given by gcc (so they are in the installation path of gcc)
and the two later have to be written with the kernel.