Add the classes List and Allocator
The List class is a linked list, and the Allocator class is the default
implementation of the template parameter AlLocatorType.
This template parameter allow the user to give a custom allocator,
such as one that allocate elements on a specific heap (this is a use
case of the class List by the kernel).