Make the ListNodeAllocator member variables static
In this way, the members are not initialized on the stack.
This is useful when initializing the free and allocated lists,
by avoid having tons of BitSet and arrays on the stack, that may result
in a stack overflow.

But the counterparts is that the MemoryManager is now static, making it
unusable for any other purpose that the kernel heap management.
This should be fixed.