Rename 2 classes in ps2 namespace
KeyboardMapper has been renamed to ScancodeMapper, and ScancodeSequence
to Scancode.

The first renaming is intented to not confuse the code when a class that
map key to ASCII character will be written. ScancodeMapper is closely
related to scancodes and is low-level, and a potential new class
KeyboardMapper will be related to KeyEvent and characters, which is
higher level.

The second renaming is because of a misunderstanding of what is a
scancode. I thought a scan code is a byte, and a sequence of scancode is
triggered when a key is pressed. But a scancode is a sequence of bytes,
and a unique scancode is triggered for each pressed/released key.