Add the Kernel class and change cpp ISRs' location
The class kernel is used to make all inportant objects (drivers,
memory managers, ...) availables to the ISRs. So now these objects are
members of the Kernel instance.

A little of hack had to be done in order to constructs the kernel
statically, but with the address of the multiboot info structure. Now
the boot assembly routine set a global variable with this address before
the static initialization of other global objects, rather than passing
it as argument to the main function.