Moves ISRs declarations inside a dedicated class
There was a problem when the ISRs were declared in the Kernel class: as
soon as something in Kernel.hpp or its included headers change, all ISRs
definitions had to be recompiled. Now only the ISRs that effectively
need the kernel include it, so only a few interrupts have to be
recompiled when something related to kernel's header change.