Remove most of standard headers from include/c++
After some tries, it seems really too hard to fully copy/paste the
standard c++ library into the kernel code. So the decision has been
taken that whenever possible, if a class is needed, that the class have
to be hand-written. So that we'll learn more, and everything is
hand-made, although a bit less safe.