Add doxygen configuration file
With that file, an option to build the documentation can be added later
to the cmake files.

Since documentation is generated in the doxygen/ directory, this
directory has to be ignored (we don't want to add the doc output to the
repository), but the config file should anyway be added.