Replaces MemoryMapBrowser by PhysicalMemoryManager
This class manages physical memory frames at a low level, and has the
responsability of allocating/freeing them.

This class use a specific mechanism of stack for its internal
management, so an instance of a MemoryManager would not do the trick.
A benefit of this mechanism is that it has no memory footprint, since
free frames are used themselves as data structure.