Use a bitset in PhysicalMemoryManager instead of a list
This solves an issue that arises when enabling paging: the physical
memory manager have to access arbitrarily pages for its management, but
since the paging is enabled, some of them are not accessible because
they are not in virtual memory.

One solution is to identity map all pages for the physical memory
manager, but it seems to be a bit overkill, and it breaks the separation
between the physical and the virtual memory management.

On the other hand, we can manage the physical memory in such way that
the manager has not to access the pages. This is what we do with a
bitset.