Add test code for paging initialization
This code is not intended to be a feature of the kernel, but it is there to help
debugging the paging initialization. Although this kind of code should not
normally appear in the repo tree, the tracked bug is so hard to solve that it is
a pain to erase the testing code before making a commit on an unrelated topic.

(Yeah I should do a branch, I know).