Fix bug in physical memory manager
DAMN GOD DAMMIT I FUCKING FOUND IT!!!!!!!!!!!!!!!!!
This bug was caused by the bitset allocated by the PMM that allocated a bitset
with a bit for each addressable byte, rather than one bit for one page.
In consequence, the bitset was huuuuge, and caused out of memory access when
paging was enabled, because the number of paged frames was too low.