Super-dirty workaround: allocate Kernel in a char array
There is an issue that when the kernel object is initialized, the
KeyboardDriver object is constructed, which asks for a dynamically
allocated list object. But since the kernel is not yet constructed, the
static attribute _kernel points to nullptr, therefore the operator new
uses nullptr as kernel instance to whom ask for memory. Bam, crash.

The true fix is to use fixed stack memory rather than dynamic one in
KeyboardDriver. That require the implementation of a ring buffer, which
will be done asap.