I think we don't need __cxa_atexit stuff
The function __cxa_atexit should be called whenever an object is
constructed, and it is used to register constructors to be called at the
exit of the kernel. But some indices led me to think that it may not
apply to the ABI used by i386-elf-g++, although it is not clear at all.
Anyway, what is clear is that this function is never called at startup
of the kernel, even though some global object constructors are being
called. So I removed all this code, it is still in the history if we
ever need it, and on OSDev.org.